+++ CHANGELOG.md
@@ -2,6 +2,13 @@
 
 ## Unreleased
 
+### Breaking Changes
+
+#### Clients & Protocols
+
+- `XdgPositionerState` moved to `XdgPopupState` and added to `XdgRequest::NewPopup`
+- `PopupSurface::send_configure` now checks the protocol version and returns an `Result`
+
 ### Additions
 
 #### Clients & Protocols
@@ -13,6 +20,7 @@
 - Whether a surface is toplevel equivalent can be determined with the new function `shell::is_toplevel_equivalent`.
 - Setting the parent of a toplevel surface is now possible with the `xdg::ToplevelSurface::set_parent` function.
 - Add support for the zxdg-foreign-v2 protocol.
+- Support for `xdg_wm_base` protocol version 3
 
 ### Bugfixes